feat(api): check for user permissions in update profile image

+++ api/src/features/profile-picture/profile-picture.controller.ts
@@ -33,24 +33,29 @@ export class ProfilePictureController {
     const quality = 50
     const { userId } = ctx.params
 
-    const { path } = ctx.state.files.image
-    const imgName = `${userId}.jpg`
+    if (ctx.state.jwt.userId === userId) {
+      const { path } = ctx.state.files.image
+      const imgName = `${userId}.jpg`
+  
+      let buf: Buffer = await sharp(path)
+        .resize(640, 640)
+        .jpeg({ quality })
+        .toBuffer()
+      MinioClient.store(bucket, region, `640x640/${imgName}`, buf)
+  
+      buf = await sharp(path).resize(160, 160).jpeg({ quality }).toBuffer()
+      MinioClient.store(bucket, region, `160x160/${imgName}`, buf)
+  
+      fs.unlink(path, (err) => {
+        if (err) throw err
+        Logger.info(`File deleted ${path}`)
+      })
+  
+      ctx.status = statusCode.OK
+      return
+    }
+    ctx.status = statusCode.FORBIDDEN
 
-    let buf: Buffer = await sharp(path)
-      .resize(640, 640)
-      .jpeg({ quality })
-      .toBuffer()
-    MinioClient.store(bucket, region, `640x640/${imgName}`, buf)
-
-    buf = await sharp(path).resize(160, 160).jpeg({ quality }).toBuffer()
-    MinioClient.store(bucket, region, `160x160/${imgName}`, buf)
-
-    fs.unlink(path, (err) => {
-      if (err) throw err
-      Logger.info(`File deleted ${path}`)
-    })
-
-    ctx.status = statusCode.OK
   }
 
   public async remove(ctx: Koa.DefaultContext, next: Function): Promise<void> {
